+++ .expo/settings.json
@@ -5,5 +5,6 @@
   "minify": false,
   "urlRandomness": null,
   "https": false,
-  "scheme": null
+  "scheme": null,
+  "devClient": false
 }
